FU-107 feat: ApiResponseDto, ProductService Interface 삭제
