--- src/main/java/com/foru/freebe/profile/entity/ApiResponseDto.java
@@ -1,21 +1,0 @@
-package com.foru.freebe.profile.entity;
-
-import org.springframework.http.HttpStatus;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonInclude(value = JsonInclude.Include.NON_NULL)
-public class ApiResponseDto<T> {
-	private HttpStatus status;
-	private String message;
-	private T data;
-}
